+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `./src/index.js` to support `import`
 - Added a README and CHANGELOG
 - Added a placeholder HelloWorld component
+<<<<<<< HEAD
 - Added @nylira/vue-page-header as `tmPageHeader`
 - Added @nylira/vue-notifications as `tmNotifications`
 - Added @nylira/vue-countdown as `tmCountdown`
@@ -19,3 +20,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added @nylira/vue-field as `tmField`
 
 ### Changes
+=======
+- Added @nylira/vue-module(s) as `tmModule(s)`
+
+### Changes
+
+- Added @nylira/vue-field as `tmField`
+>>>>>>> 171143776c032c4c9555c37f5bfae49e10bfe587